GeminiHelper/requirements.txt
google-generativeai
watchdog
